pyproject.toml
[tool.poetry]
name = "python-api-automation-framework"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "**/*" }]

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.4.0"
requests = "^2.31.0"
toml = "^0.10.2"
pytest = "^8.1.1"
pydantic = "^2.7.0"
allure-pytest = "^2.13.5"
pytest-xdist = "^3.6.1"
jsonpath-ng = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"